requirements.txt
aiohttp
numpy
Pillow
pytesseract
scipy
ultralytics
watchdog
